refactored: Общий рефакторинг
