+++ src/pages/PublicPage.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import MainInfo from "@/pages/publicPage/components/MainInfo.vue";
-import LineDivider from "./publicPage/components/LineDivider.vue";
+import LineDivider from "../components/common/LineDivider.vue";
 import ContactInfo from "./publicPage/components/ContactInfo.vue";
 </script>
