--- src/pages/publicPage/components/LineDivider.vue
@@ -1,10 +1,0 @@
-<template>
-  <div class="line-divider" />
-</template>
-
-<style scoped lang="scss">
-.line-divider {
-  padding: 10px 0;
-  border-bottom: 1px solid $gray-color;
-}
-</style>
